Discriminate query plan caches by the assembly depending return type
When having multiple assemblies with the same type (this can happen when
dynamically generating assemblies or loading assemblies multiple times),
NHibernate used to return query plan cache entries with mismatching types.

The query portion of the cache key contained the type, but not the
assembly. Thus two types from different assemblies, but with the same
name lead to returning plans with wrong return types.